chore(storybook): clean masthead and dotcom stories
