+++ packages/react/src/components/Masthead/__stories__/Masthead.stories.js
@@ -2,47 +2,9 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select, boolean } from '@storybook/addon-knobs';
 import Masthead from '../Masthead';
-import mastheadLinks from './data/MastheadLinks.js';
+import mastheadKnobs from './data/Masthead.stories.knobs.js';
 import readme from '../README.md';
 import '../../../../../styles/scss/components/masthead/index.scss';
-
-const platformName = {
-  none: 'undefined',
-  platform: {
-    name: 'IBM Cloud',
-    url: 'https://www.ibm.com/cloud',
-  },
-};
-
-const props = {
-  navigation: () => {
-    return {
-      navigation: select('Navigation', {
-        default: 'string',
-        custom: mastheadLinks,
-        none: false,
-      }),
-    };
-  },
-  platform: () => {
-    return {
-      platform: select('Platform', {
-        none: 'undefined',
-        hasName: platformName.platform,
-      }),
-    };
-  },
-  profile: () => {
-    return {
-      hasProfile: boolean('Show profile', true),
-    };
-  },
-  search: () => {
-    return {
-      hasSearch: boolean('Show search', true),
-    };
-  },
-};
 
 storiesOf('Masthead', module)
   .addDecorator(withKnobs)
@@ -54,10 +16,18 @@ storiesOf('Masthead', module)
   .add('Default', () => {
     return (
       <Masthead
-        {...props.navigation()}
-        {...props.platform()}
-        {...props.profile()}
-        {...props.search()}
+        navigation={select(
+          'Navigation',
+          mastheadKnobs.navigation,
+          mastheadKnobs.navigation.default
+        )}
+        platform={select(
+          'Platform name',
+          mastheadKnobs.platform,
+          mastheadKnobs.platform.none
+        )}
+        hasProfile={boolean('Show profile', true)}
+        hasSearch={boolean('Show search', true)}
       />
     );
   });
